Split BaseHelper off into MessageHelper
